Fix java compiler warnings for HCS plugin
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@12281 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ plugins/HCS/src/org/micromanager/hcs/PlatePanel.java
@@ -373,8 +373,8 @@ public class PlatePanel extends JPanel {
       activeRect_ = getBounds();
 
       // shrink drawing area by the margin amount
-      activeRect_.x = (int)xMargin_;
-      activeRect_.y = (int)yMargin_;
+      activeRect_.x = xMargin_;
+      activeRect_.y = yMargin_;
       activeRect_.height -= 2*yMargin_;
       activeRect_.width -= 2*xMargin_;
